refactor: Add Discord key authentication to whitelist and signup routes

+++ src/app/api/survey/[id]/route.ts
@@ -2,14 +2,28 @@ import connectMongoDB from "@/lib/mongodb";
 import Survey from "@/models/surveys";
 import { NextResponse } from "next/server";
 import { ZodError } from "zod";
+import { headers } from "next/headers";
+import { getServerSession } from "next-auth";
+import { authOptions } from "@/lib/auth";
 
 export async function GET(
   request: Request,
   { params: { id } }: { params: { id: string } }
 ) {
+  const session = await getServerSession(authOptions);
+  const discordKey = headers().get("discord-key");
+  const user_id = session?.user?.id;
+
   try {
     await connectMongoDB();
     const survey = await Survey.findOne({ discordId: id });
+    if (
+      !session ||
+      !discordKey ||
+      discordKey !== process.env.BOT_KEY ||
+      user_id !== survey.discordId
+    )
+      return new Response("401");
     return NextResponse.json({ survey }, { status: 200 });
   } catch (error) {
     if (error instanceof ZodError)
@@ -23,6 +37,11 @@ export async function DELETE(
   request: Request,
   { params: { id } }: { params: { id: string } }
 ) {
+  const discordKey = headers().get("discord-key");
+
+  if (!discordKey || discordKey !== process.env.BOT_KEY)
+    return new Response("401");
+
   await connectMongoDB();
   await Survey.findByIdAndDelete(id);
   return NextResponse.json({ message: "Survey deleted" }, { status: 200 });
